Create a PostsClient class with getPosts and addNewPost
