+++ src/client/types.ts
@@ -2,5 +2,5 @@ import { Post, PostDataWithoutId } from "../posts/types";
 
 export interface PostClientstructure {
   getPost: () => Promise<Post[]>;
-  newPost: (postData: PostDataWithoutId) => Promise<Post>;
+  addNewPost: (postData: PostDataWithoutId) => Promise<Post>;
 }
